refactor!: switch to using Firestore’s native Codable support
BREAKING CHANGE: Added FirebaseFirestoreSwift (part of the Firebase SDK) for native Codable support. This replaces the FirebaseConvertable protocol and adds support for optional properties and URLs, meaning that the workaround we use to store image URLs now is no longer needed. This is a breaking change because post timestamps are now encoded differently and the types are not compatible.